refactor(add-recipe): extract useRecipeForm hook (useReducer-backed) (R1)
Move all create/edit-recipe form logic out of the 579-line AddRecipe god
component into a useRecipeForm hook. Field state is now a typed useReducer
(SET_FIELD supports the useState updater-fn form, so child setters keep their
React.Dispatch<SetStateAction<T>> contract — IngredientsContainer /
InstructionsContainer rely on functional updates); ancillary UI/status state
(loading, errors, draft bookkeeping) stays as useState. The hook owns the draft
hydration + URL-sync + validation effects, the draft-autosave wiring, and
handleSubmit; AddRecipe.tsx is now purely presentational (~200 lines, JSX only).

Behaviour-preserving: no field/effect/submit logic changed, only relocated.
Full Vitest suite green (576 pass / 2 skip) incl. the 25-case AddRecipe suite
and the draft-resume flows.
