+++ src/pages/AddRecipe/AddRecipe.tsx
@@ -1,18 +1,7 @@
 import { InfoIcon } from 'src/Components/icons'
-import React, { FC, useEffect, useMemo, useRef, useState } from 'react'
-import axios from 'axios'
-import { useNavigate, useSearchParams } from 'react-router-dom'
-import { useQueryClient } from '@tanstack/react-query'
-import {
-  AddRecipeErrorType,
-  IngredientsType,
-  InstructionsType,
-  RecipeDraftContent,
-  RecipeDraftType,
-  RecipeEditFormType,
-  RecipeFormType,
-  RecipeType,
-} from 'types'
+import React, { FC } from 'react'
+import { useNavigate } from 'react-router-dom'
+import { RecipeType } from 'types'
 import LoadingBar from 'react-top-loading-bar'
 import FormInput from 'src/Components/Form/FormInput'
 import ImagePicker from 'src/pages/AddRecipe/ImagePicker/ImagePicker'
@@ -25,355 +14,62 @@ import InstructionsContainer from 'src/pages/AddRecipe/Instructions/Instructions
 import CuisineSelector from 'src/pages/AddRecipe/CuisineSelector/CuisineSelector'
 import MealTypeSelector from 'src/pages/AddRecipe/MealTypeSelector/MealTypeSelector'
 import DietSelector from 'src/pages/AddRecipe/DietSelector/DietSelector'
-import { hrMinToMin } from 'src/util/hrMinToMin'
-import { minToHrMin } from 'src/util/minToHrMin'
 import { TITLE_MAX_LENGTH, DESCRIPTION_MAX_LENGTH } from 'src/util/recipeLimits'
-import {
-  validateRecipeForm,
-  isRecipeFormValid,
-} from 'src/pages/AddRecipe/recipeFormValidation'
-import RecipeAPI from 'src/api/recipes'
 import styles from 'src/_exports.module.scss'
 import AddRecipeFormError from 'src/pages/AddRecipe/AddRecipeFormError'
 import SectionHeader from 'src/pages/AddRecipe/SectionHeader'
 import AddRecipeSummaryBar from 'src/pages/AddRecipe/AddRecipeSummaryBar'
 import { Helmet } from 'react-helmet-async'
-import { toast } from 'react-hot-toast'
-import DraftAPI from 'src/api/drafts'
-import { useDraftAutosave } from 'src/pages/AddRecipe/useDraftAutosave'
 import DraftSaveStatus from 'src/pages/AddRecipe/DraftSaveStatus'
 import DraftResumeBanner from 'src/pages/AddRecipe/DraftResumeBanner'
-
-type TimeVal = { hours: number; minutes: number } | null
-
-// Same auth-expiry message on both the update and publish paths below.
-const SESSION_EXPIRED =
-  'Your session has expired — please sign in again and retry.'
-
-// Shown when a recipe saves but is held by automated moderation for an admin to
-// review before it appears publicly. Longer-lived than a normal toast (and not
-// styled as an error — nothing went wrong) so the owner doesn't miss it.
-const notifyPendingReview = () =>
-  toast(
-    'Your recipe was submitted and is pending review. It will appear publicly once approved.',
-    { icon: '⏳', duration: 7000 }
-  )
+import { useRecipeForm } from 'src/pages/AddRecipe/useRecipeForm'
 
 // When `initialRecipe` is supplied the form runs in edit mode: every field is
 // pre-populated from the existing recipe and submitting updates it (preserving
-// ratings/saves) instead of creating a new one.
+// ratings/saves) instead of creating a new one. All form state, validation,
+// draft autosave and submit orchestration live in useRecipeForm; this component
+// is the presentation layer.
 type AddRecipeProps = { initialRecipe?: RecipeType }
 
 const AddRecipe: FC<AddRecipeProps> = ({ initialRecipe }) => {
-  const isEditMode = !!initialRecipe
-
-  const [addRecipeLoading, setAddRecipeLoading] = useState(false)
-  const [loadingProgress, setLoadingProgress] = useState(0)
-  const addRecipeFormRef = useRef<HTMLDivElement>(null)
-  const [title, setTitle] = useState(initialRecipe?.title ?? '')
-  const [recipeImage, setRecipeImage] = useState<File | undefined>()
-  // Edit mode only: the recipe's current image URL, kept when the user doesn't
-  // pick a new file. Cleared if they remove the image (forcing a new pick).
-  const [existingImageUrl, setExistingImageUrl] = useState<string | undefined>(
-    initialRecipe?.recipeImage
-  )
-  const [description, setDescription] = useState(initialRecipe?.description ?? '')
-  const [servings, setServings] = useState<number | ''>(
-    initialRecipe?.servings ?? ''
-  )
-  const [prepTime, setPrepTime] = useState<TimeVal>(
-    initialRecipe ? minToHrMin(initialRecipe.prepTime) : null
-  )
-  const [cookTime, setCookTime] = useState<TimeVal>(
-    initialRecipe ? minToHrMin(initialRecipe.cookTime) : null
-  )
-  const [fridgeLife, setFridgeLife] = useState<number>(
-    initialRecipe?.fridgeLife ?? 0
-  )
-  const [freezerLife, setFreezerLife] = useState<number>(
-    initialRecipe?.freezerLife ?? 0
-  )
-  const [ingredients, setIngredients] = useState<IngredientsType[]>(
-    initialRecipe?.ingredients ?? []
-  )
-  const [instructions, setInstructions] = useState<InstructionsType[]>(
-    initialRecipe?.instructions ?? []
-  )
-  const [cuisine, setCuisine] = useState(initialRecipe?.cuisine ?? '')
-  const [mealTypes, setMealTypes] = useState<string[]>(
-    initialRecipe?.mealTypes ?? []
-  )
-  // Author-selected diet tags (optional). In edit mode this pre-fills from the
-  // recipe's existing nutritionLabels — including ones Edamam set on older
-  // recipes — so the author can keep or correct them.
-  const [nutritionLabels, setNutritionLabels] = useState<string[]>(
-    initialRecipe?.nutritionLabels ?? []
-  )
-  const [errors, setErrors] = useState<Partial<AddRecipeErrorType>>({})
-
-  const [isFormValid, setIsFormValid] = useState(false)
-  const [hasAttemptedSubmit, setHasAttemptedSubmit] = useState(false)
-
   const navigate = useNavigate()
-  const queryClient = useQueryClient()
-
-  // ─── Draft autosave (create flow only) ──────────────────────────────────────
-  // Edit mode never touches drafts. In create mode the form is autosaved to a
-  // server-side draft so unfinished recipes survive a refresh or navigating
-  // away. The image is intentionally not part of a draft — it's re-picked on
-  // resume (publish validation still requires one).
-  const [searchParams, setSearchParams] = useSearchParams()
-  const urlDraftId = isEditMode ? null : searchParams.get('draftId')
-  const [draftId, setDraftId] = useState<string | null>(urlDraftId)
-  // "Hydrated" gates autosave: true for a fresh create, briefly false while a
-  // resumed draft loads into the fields.
-  const [hydrated, setHydrated] = useState(!urlDraftId)
-  // Draft ids whose content is already on screen — either just loaded here, or
-  // just created by autosave. The hydration effect skips these so our own URL
-  // updates (and a resume of the draft we're already editing) don't trigger a
-  // redundant reload.
-  const ownedDraftsRef = useRef<Set<string>>(new Set())
-  // True once a draft was resumed (loaded from the server) this session, used to
-  // remind the user to re-add the image — drafts don't persist it. Not set for
-  // drafts created in the current session (the user never had an image to lose).
-  const [resumedFromDraft, setResumedFromDraft] = useState(false)
-
-  // Load the draft named in the URL whenever it points at one we haven't loaded
-  // yet. Runs on mount for `?draftId=…`, and again when the resume banner
-  // navigates to a draft while this page is already mounted — same route, so no
-  // remount happens on its own and a mount-only effect would never re-fire.
-  useEffect(() => {
-    if (isEditMode || !urlDraftId || ownedDraftsRef.current.has(urlDraftId)) return
-    let cancelled = false
-    setHydrated(false)
-    DraftAPI.getDraft(urlDraftId)
-      .then(draft => {
-        if (cancelled) return
-        ownedDraftsRef.current.add(urlDraftId)
-        if (draft) {
-          setDraftId(urlDraftId)
-          setTitle(draft.title ?? '')
-          setDescription(draft.description ?? '')
-          setServings(draft.servings ?? '')
-          setPrepTime(draft.prepTime != null ? minToHrMin(draft.prepTime) : null)
-          setCookTime(draft.cookTime != null ? minToHrMin(draft.cookTime) : null)
-          setFridgeLife(draft.fridgeLife ?? 0)
-          setFreezerLife(draft.freezerLife ?? 0)
-          setIngredients(draft.ingredients ?? [])
-          setInstructions(draft.instructions ?? [])
-          setCuisine(draft.cuisine ?? '')
-          setMealTypes(draft.mealTypes ?? [])
-          setNutritionLabels(draft.nutritionLabels ?? [])
-          setResumedFromDraft(true)
-        }
-        setHydrated(true)
-      })
-      .catch((err: unknown) => {
-        if (cancelled) return
-        const status = axios.isAxiosError(err) ? err.response?.status : undefined
-        if (status === 404 || status === 403) {
-          // The draft is genuinely gone or not the caller's. Drop only the
-          // draftId param (preserving any other query params) and let the user
-          // start a fresh draft.
-          toast.error("Couldn't load that draft — starting a new one.")
-          setDraftId(null)
-          const next = new URLSearchParams(searchParams)
-          next.delete('draftId')
-          setSearchParams(next, { replace: true })
-          setHydrated(true)
-        } else {
-          // Transient failure (network/5xx) on a draft that likely still
-          // exists. Leave autosave disabled (hydrated stays false) so we don't
-          // create a duplicate or overwrite the unloaded draft with a partial
-          // form; ask the user to retry.
-          toast.error('Could not load your draft. Refresh to try again.')
-        }
-      })
-    return () => {
-      cancelled = true
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [urlDraftId])
-
-  // Serializable draft content mirrored from form state. Times are stored as
-  // minutes (matching the recipe shape); empty values are omitted.
-  const draftContent: RecipeDraftContent = useMemo(
-    () => ({
-      title,
-      description,
-      // Send explicit null (not undefined) when empty so clearing a field is
-      // persisted rather than silently dropped from the payload — see
-      // RecipeDraftContent.
-      servings: servings === '' ? null : Number(servings),
-      prepTime: prepTime ? hrMinToMin(prepTime) : null,
-      cookTime: cookTime ? hrMinToMin(cookTime) : null,
-      fridgeLife,
-      freezerLife,
-      ingredients,
-      instructions,
-      cuisine,
-      mealTypes,
-      nutritionLabels,
-    }),
-    [
-      title,
-      description,
-      servings,
-      prepTime,
-      cookTime,
-      fridgeLife,
-      freezerLife,
-      ingredients,
-      instructions,
-      cuisine,
-      mealTypes,
-      nutritionLabels,
-    ]
-  )
-
-  // A brand-new draft is only created once the recipe has a title. This gates
-  // out throwaway drafts from a stray keystroke (keeping the Drafts list and the
-  // per-user draft count clean); updating an existing draft is unaffected.
-  const canCreateDraft = !!title.trim()
-
-  const { status: draftStatus, clearDraft } = useDraftAutosave({
-    content: draftContent,
-    enabled: !isEditMode && hydrated,
-    canCreate: canCreateDraft,
-    draftId,
-    onDraftCreated: (draft: RecipeDraftType) => {
-      // Mark as owned before the URL sync below points the URL at it, so the
-      // hydration effect doesn't reload the draft we just created.
-      ownedDraftsRef.current.add(draft._id)
-      setDraftId(draft._id)
-      queryClient.invalidateQueries({ queryKey: ['drafts'] })
-    },
-    onLimitReached: (message: string) => toast.error(message),
-  })
-
-  // Reflect the active draft id in the URL (replace) so a refresh resumes the
-  // same draft rather than starting a second one.
-  useEffect(() => {
-    if (isEditMode || !draftId) return
-    if (searchParams.get('draftId') === draftId) return
-    const next = new URLSearchParams(searchParams)
-    next.set('draftId', draftId)
-    setSearchParams(next, { replace: true })
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [draftId])
-
-  const validate = (assignErrors: boolean = false) => {
-    const newErrors = validateRecipeForm({
-      title,
-      recipeImage,
-      existingImageUrl,
-      description,
-      servings,
-      prepTime,
-      ingredients,
-      instructions,
-      mealTypes,
-    })
-    assignErrors && setErrors(newErrors)
-    return isRecipeFormValid(newErrors)
-  }
-  useEffect(() => {
-    // Once the user has attempted a submit, keep the displayed errors in sync as
-    // fields are fixed (assignErrors=true) so a corrected field clears its message
-    // immediately instead of lingering until the next submit click. Before the
-    // first attempt we only compute validity, never surface errors.
-    if (validate(hasAttemptedSubmit)) setIsFormValid(true)
-    else setIsFormValid(false)
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [
+  const {
+    isEditMode,
     title,
     recipeImage,
     existingImageUrl,
     description,
     servings,
     prepTime,
+    cookTime,
     ingredients,
     instructions,
+    cuisine,
     mealTypes,
-    hasAttemptedSubmit,
-  ])
-  const handleSubmit = async () => {
-    if (addRecipeLoading) return
-    setHasAttemptedSubmit(true)
-    if (!validate(true)) {
-      addRecipeFormRef?.current && addRecipeFormRef.current.scrollTo(0, 0)
-      return
-    }
-
-    setAddRecipeLoading(true)
-    // Shared form-state → payload mapping; the two branches differ only in the
-    // image field (optional on edit, required on create) and which API they call.
-    const formData = {
-      title,
-      prepTime: hrMinToMin(prepTime),
-      cookTime: hrMinToMin(cookTime),
-      servings: Number(servings),
-      fridgeLife,
-      freezerLife,
-      description,
-      ingredients,
-      instructions,
-      cuisine,
-      mealTypes,
-      nutritionLabels,
-    }
-    if (isEditMode && initialRecipe) {
-      const editData: RecipeEditFormType = { ...formData, recipeImage }
-      const result = await RecipeAPI.editRecipe(
-        initialRecipe._id,
-        editData,
-        initialRecipe,
-        setLoadingProgress
-      )
-      if (result.status === 'auth-error') {
-        toast.error(SESSION_EXPIRED)
-      } else if (result.status === 'success') {
-        // Write the server's updated recipe straight into the cache rather than
-        // invalidating: invalidation would refetch GET /getRecipe (which bumps
-        // the public view counter) and briefly flash stale data. The created
-        // list still needs a refetch to reorder/relabel.
-        queryClient.setQueryData(['recipe', initialRecipe._id], result.recipe)
-        queryClient.invalidateQueries({ queryKey: ['created-recipes'] })
-        // A medium-confidence moderation hold saves the edit but withholds it from
-        // public reads until an admin clears it; tell the owner instead of the
-        // usual "updated" so a silently-hidden recipe isn't a surprise.
-        if (result.recipe.status === 'pending_review') {
-          notifyPendingReview()
-        } else {
-          toast.success('Recipe updated.')
-        }
-        navigate(`/recipes/${initialRecipe._id}`)
-      } else {
-        toast.error(result.message)
-      }
-    } else {
-      const recipeData: RecipeFormType = { ...formData, recipeImage: recipeImage! }
-      const result = await RecipeAPI.addRecipe(recipeData, setLoadingProgress)
-      if (result.status === 'auth-error') {
-        toast.error(SESSION_EXPIRED)
-      } else if (result.status === 'success') {
-        // Recipe is saved — remove the now-redundant draft (and stop autosave
-        // from recreating it on unmount) before navigating away.
-        await clearDraft()
-        queryClient.invalidateQueries({ queryKey: ['drafts'] })
-        if (result.pendingReview) {
-          notifyPendingReview()
-        } else {
-          toast.success('Recipe published.')
-        }
-        navigate(`/recipes/${result.id}`)
-      } else {
-        toast.error(result.message)
-      }
-    }
-    setAddRecipeLoading(false)
-    setLoadingProgress(100)
-  }
+    nutritionLabels,
+    setTitle,
+    setRecipeImage,
+    setExistingImageUrl,
+    setDescription,
+    setServings,
+    setPrepTime,
+    setCookTime,
+    setIngredients,
+    setInstructions,
+    setCuisine,
+    setMealTypes,
+    setNutritionLabels,
+    errors,
+    isFormValid,
+    addRecipeLoading,
+    loadingProgress,
+    setLoadingProgress,
+    draftStatus,
+    draftId,
+    resumedFromDraft,
+    addRecipeFormRef,
+    handleSubmit,
+  } = useRecipeForm(initialRecipe)
 
   return (
     <>
